+++ HISTORY.md
@@ -4,4 +4,4 @@
 
 0.0.1 / 2015-03-21
 ==================
-  * Prowl created
+  * Blurr created